Fixed issue 31
Gui now hides the EOF flag from the user. This allows the user to write freely, and UVSim just adds an eof flag internally. Also added a check in memory to see if the address in the operation is outside of the defined program. This fixed weird behavior in Test3b.